Remove lib_deity. Remaining funcs in lib_helpers
